fix avt, background is null os empty

+++ src/main/java/com/example/socialnetwork/domain/service/UserServiceImpl.java
@@ -132,13 +132,13 @@ public class UserServiceImpl implements UserServicePort {
         if(!isDeleteAvt){
             String avt = HandleFile.loadFileImage(profileRequest.getAvatar(), storageServicePort, 1);
             if(avt != null){
-                if(user.getAvatar() != null){
+                if(user.getAvatar() != null && !user.getAvatar().isEmpty()){
                     s3Service.deleteFile(HandleFile.getFilePath(user.getAvatar()));
                 }
                 user.setAvatar(avt);
             }
         }else{
-            if(user.getAvatar() != null){
+            if(user.getAvatar() != null && !user.getAvatar().isEmpty()){
                 s3Service.deleteFile(HandleFile.getFilePath(user.getAvatar()));
                 user.setAvatar(null);
             }
@@ -147,13 +147,13 @@ public class UserServiceImpl implements UserServicePort {
         if(!isDeleteBackground){
             String background = HandleFile.loadFileImage(profileRequest.getBackground(), storageServicePort, 1);
             if(background != null){
-                if(user.getBackgroundImage() != null){
+                if(user.getBackgroundImage() != null && !user.getBackgroundImage().isEmpty()){
                     s3Service.deleteFile(HandleFile.getFilePath(user.getBackgroundImage()));
                 }
                 user.setBackgroundImage(background);
             }
         }else{
-            if(user.getBackgroundImage() != null){
+            if(user.getBackgroundImage() != null && !user.getBackgroundImage().isEmpty()){
                 s3Service.deleteFile(HandleFile.getFilePath(user.getBackgroundImage()));
                 user.setBackgroundImage(null);
             }
